+++ src/providers/executeCommandCompletionProvider.ts
@@ -15,18 +15,23 @@ export function registerExecuteCommandCompletionProvider(context: ExtensionConte
   );
 }
 
+interface CommandMetadata {
+  command: string;
+  extensionName?: string;
+  title?: string;
+}
+
 /**
  * Provide autocompletion for `executeCommand`.
  */
 export class ExecuteCommandCompletionProvider
-  implements vscode.Disposable, vscode.CompletionItemProvider
-{
-  private readonly commandIds: Lazy<Thenable<string[]>>;
+  implements vscode.Disposable, vscode.CompletionItemProvider {
+  private readonly commandMetadata: Lazy<Promise<readonly CommandMetadata[]>>;
   private readonly disposables: vscode.Disposable[];
 
   constructor() {
-    this.commandIds = new Lazy(() => vscode.commands.getCommands(true));
-    this.disposables = [vscode.extensions.onDidChange(() => this.commandIds.reset())];
+    this.commandMetadata = new Lazy(() => this.loadCommandMetadata());
+    this.disposables = [vscode.extensions.onDidChange(() => this.commandMetadata.reset())];
   }
 
   dispose() {
@@ -47,17 +52,37 @@ export class ExecuteCommandCompletionProvider
       return;
     }
 
-    const commandIds = await this.commandIds.get();
     const quote = match.groups?.quote1 || match.groups?.quote2;
     const range = document.getWordRangeAtPosition(position, /[a-zA-Z0-9._-]+/);
 
-    return commandIds.map((id) => {
+    const commandMetadata = await this.commandMetadata.get();
+    return commandMetadata.map(({ command: id, extensionName, title }) => {
       const item = new vscode.CompletionItem(id, vscode.CompletionItemKind.Value);
+      item.detail = extensionName ? `${title ?? ''} [${extensionName}]` : title;
       item.range = range;
       if (!quote) {
         item.insertText = `'${id}'`;
       }
       return item;
     });
+  }
+
+  private async loadCommandMetadata(): Promise<readonly CommandMetadata[]> {
+    const cmd2Metadata = new Map<string, CommandMetadata>();
+    for (const { packageJSON } of vscode.extensions.all) {
+      const cmds: CommandMetadata[] = packageJSON?.contributes?.commands;
+      if (cmds) {
+        const extensionName: string | undefined = packageJSON.displayName ?? packageJSON.name;
+        cmds.forEach((m) => cmd2Metadata.set(m.command, { ...m, extensionName }));
+      }
+    }
+
+    const cmdMetadata: CommandMetadata[] = [];
+    for (const command of await vscode.commands.getCommands(true)) {
+      const metadata = cmd2Metadata.get(command);
+      cmdMetadata.push(metadata ?? { command });
+    }
+
+    return Object.freeze(cmdMetadata);
   }
 }